requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic[dotenv]
sqlalchemy
alembic
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
requests
swarms
groq
python-dotenv
pandas
openpyxl
PyPDF2
pydantic
Pillow
pdf2image
pdfplumber
google-generativeai
scikit-learn
faker
orjson
swarm-models
swarm_models
langchain_community